refactor: migrate queries and actions to server directory
- Moved various queries and actions from the lib directory to the server directory for improved organization and clarity.
- Updated import paths across multiple components to reflect the new structure.
- Removed obsolete files related to barometers, brands, categories, conditions, and reports to streamline the codebase.
