+++ app/admin/add-barometer/barometer-form.schema.ts
