+++ app/admin/add-brand/brand-add-form.tsx
@@ -8,10 +8,11 @@ import { z } from 'zod'
 import { IconUpload, ImageUpload, MultiSelect, RequiredFieldMark } from '@/components/elements'
 import * as UI from '@/components/ui'
 import { imageStorage } from '@/constants'
-import { createBrand } from '@/lib/brands/actions'
-import type { AllBrandsDTO } from '@/lib/brands/queries'
-import type { CountryListDTO } from '@/lib/counties/queries'
-import { createImageUrls, uploadFileToCloud } from '@/lib/images/actions'
+import { createBrand } from '@/server/brands/actions'
+import type { AllBrandsDTO } from '@/server/brands/queries'
+import type { CountryListDTO } from '@/server/counties/queries'
+import { createImageUrls } from '@/server/images/actions'
+import { uploadFileToCloud } from '@/server/images/upload'
 import { generateIcon, getThumbnailBase64 } from '@/utils'
 
 // Zod validation schema
