+++ app/admin/add-document/document-form.tsx
@@ -10,9 +10,9 @@ import { z } from 'zod'
 import { MultiSelect, RequiredFieldMark } from '@/components/elements'
 import * as UI from '@/components/ui'
 import { imageStorage } from '@/constants/globals'
-import type { AllBarometersDTO } from '@/lib/barometers/queries'
-import type { ConditionsDTO } from '@/lib/conditions/queries'
-import { createDocument } from '@/lib/documents/actions'
+import type { AllBarometersDTO } from '@/server/barometers/queries'
+import type { ConditionsDTO } from '@/server/conditions/queries'
+import { createDocument } from '@/server/documents/actions'
 import { getThumbnailBase64 } from '@/utils'
 import { FileUpload } from '../add-barometer/file-upload'
