+++ app/brands/brand-edit.tsx
@@ -8,10 +8,10 @@ import { toast } from 'sonner'
 import { IconUpload, RequiredFieldMark } from '@/components/elements'
 import * as UI from '@/components/ui'
 import { imageStorage } from '@/constants'
-import { deleteBrand, updateBrand } from '@/lib/brands/actions'
-import type { AllBrandsDTO, BrandDTO } from '@/lib/brands/queries'
-import type { CountryListDTO } from '@/lib/counties/queries'
-import { deleteImages } from '@/lib/images/actions'
+import { deleteBrand, updateBrand } from '@/server/brands/actions'
+import type { AllBrandsDTO, BrandDTO } from '@/server/brands/queries'
+import type { CountryListDTO } from '@/server/counties/queries'
+import { deleteImages } from '@/server/images/actions'
 import { generateIcon, getThumbnailBase64 } from '@/utils'
 import { type BrandEditForm, brandEditSchema } from './brand-edit-schema'
 import { BrandImageEdit } from './brand-image-edit'
