+++ app/collection/categories/[...category]/page.tsx
@@ -8,9 +8,9 @@ import { Card, Pagination } from '@/components/ui'
 import { DEFAULT_PAGE_SIZE, imageStorage } from '@/constants'
 import { openGraph, title, twitter } from '@/constants/metadata'
 import { FrontRoutes } from '@/constants/routes-front'
-import { getBarometersByParams } from '@/lib/barometers/queries'
-import { getCategory } from '@/lib/categories/queries'
 import { withPrisma } from '@/prisma/prismaClient'
+import { getBarometersByParams } from '@/server/barometers/queries'
+import { getCategory } from '@/server/categories/queries'
 import { type DynamicOptions, SortOptions, type SortValue } from '@/types'
 import Sort from './sort'
