+++ app/collection/items/[slug]/components/carousel.tsx
@@ -4,7 +4,7 @@ import Image from 'next/image'
 import { Navigation, Pagination, Zoom } from 'swiper/modules'
 import { Swiper, SwiperSlide } from 'swiper/react'
 import { IsAdmin } from '@/components/elements'
-import type { BarometerDTO } from '@/lib/barometers/queries'
+import type { BarometerDTO } from '@/server/barometers/queries'
 import { customImageLoader } from '@/utils'
 import { ImagesEdit } from './edit-fields/images-edit'
 import 'swiper/css'
