+++ app/collection/items/[slug]/components/edit-fields/dimensions-edit.tsx
@@ -8,8 +8,8 @@ import { useFieldArray, useForm } from 'react-hook-form'
 import { toast } from 'sonner'
 import { z } from 'zod'
 import * as UI from '@/components/ui'
-import { updateBarometer } from '@/lib/barometers/actions'
-import type { BarometerDTO } from '@/lib/barometers/queries'
+import { updateBarometer } from '@/server/barometers/actions'
+import type { BarometerDTO } from '@/server/barometers/queries'
 import type { Dimensions } from '@/types'
 import { cn } from '@/utils'
