+++ app/collection/items/[slug]/components/edit-fields/edit-category.tsx
@@ -23,9 +23,9 @@ import {
   SelectTrigger,
   SelectValue,
 } from '@/components/ui'
-import { updateBarometer } from '@/lib/barometers/actions'
-import type { BarometerDTO } from '@/lib/barometers/queries'
-import type { CategoriesDTO } from '@/lib/categories/queries'
+import { updateBarometer } from '@/server/barometers/actions'
+import type { BarometerDTO } from '@/server/barometers/queries'
+import type { CategoriesDTO } from '@/server/categories/queries'
 import { EditButton } from './edit-button'
 
 interface Props {
