+++ app/collection/items/[slug]/components/edit-fields/materials-edit.tsx
@@ -8,9 +8,9 @@ import { useForm } from 'react-hook-form'
 import { toast } from 'sonner'
 import { z } from 'zod'
 import * as UI from '@/components/ui'
-import { updateBarometer } from '@/lib/barometers/actions'
-import type { BarometerDTO } from '@/lib/barometers/queries'
-import type { MaterialsDTO } from '@/lib/materials/queries'
+import { updateBarometer } from '@/server/barometers/actions'
+import type { BarometerDTO } from '@/server/barometers/queries'
+import type { MaterialsDTO } from '@/server/materials/queries'
 import { cn } from '@/utils'
 
 interface MaterialsEditProps extends ComponentProps<'button'> {
