+++ app/collection/items/[slug]/components/edit-fields/movements-edit.tsx
@@ -7,9 +7,9 @@ import { useForm } from 'react-hook-form'
 import { toast } from 'sonner'
 import { z } from 'zod'
 import * as UI from '@/components/ui'
-import { updateBarometer } from '@/lib/barometers/actions'
-import type { BarometerDTO } from '@/lib/barometers/queries'
-import type { MovementsDTO } from '@/lib/movements/queries'
+import { updateBarometer } from '@/server/barometers/actions'
+import type { BarometerDTO } from '@/server/barometers/queries'
+import type { MovementsDTO } from '@/server/movements/queries'
 import { cn } from '@/utils'
 
 interface SubcategoryEditProps extends ComponentProps<'button'> {
