+++ app/collection/items/[slug]/page.tsx
@@ -21,13 +21,13 @@ import { notFound } from 'next/navigation'
 import { IsAdmin, MD, ShowMore } from '@/components/elements'
 import { Card, SeparatorWithText } from '@/components/ui'
 import { FrontRoutes } from '@/constants'
-import { getBarometer } from '@/lib/barometers/queries'
-import { getAllBrands } from '@/lib/brands/queries'
-import { getCategories } from '@/lib/categories/queries'
-import { getConditions } from '@/lib/conditions/queries'
-import { getMaterials } from '@/lib/materials/queries'
-import { getMovements } from '@/lib/movements/queries'
 import { withPrisma } from '@/prisma/prismaClient'
+import { getBarometer } from '@/server/barometers/queries'
+import { getAllBrands } from '@/server/brands/queries'
+import { getCategories } from '@/server/categories/queries'
+import { getConditions } from '@/server/conditions/queries'
+import { getMaterials } from '@/server/materials/queries'
+import { getMovements } from '@/server/movements/queries'
 import type { Dimensions } from '@/types'
 import { BreadcrumbsComponent } from './components/breadcrumbs'
 // local components
