+++ server/README.md
@@ -1,0 +1,10 @@
+# server
+
+Server-side functions and database operations organized by domain:
+
+- **actions.ts** - Server Actions for data mutations (create, update, delete)
+- **queries.ts** - Database queries for data fetching
+- **schemas/** - Zod validation schemas
+- **search.ts** - Search functionality
+
+Each folder represents a domain entity (barometers, brands, documents, etc.) with its corresponding database operations.
