+++ server/barometers/actions.ts
