+++ server/barometers/queries.ts
