+++ server/barometers/search.ts
