+++ server/conditions/queries.ts
