+++ server/images/actions.ts
@@ -37,15 +37,4 @@ async function createImageUrls(files: FileProps[]) {
   return { urls }
 }
 
-async function uploadFileToCloud(url: string, file: File) {
-  const res = await fetch(url, {
-    method: 'PUT',
-    headers: {
-      'Content-Type': file.type,
-    },
-    body: file,
-  })
-  if (!res.ok) throw new Error(res.statusText)
-}
-
-export { deleteImages, deleteImage, createImageUrls, uploadFileToCloud }
+export { deleteImages, deleteImage, createImageUrls }
